+++ .babelrc
@@ -21,5 +21,10 @@
         "spec": true
       }
     ]
-  ]
+  ],
+  "env": {
+    "test": {
+      "presets": ["es2015", "react", "stage-0"]
+    }
+  }
 }
