+++ app/javascript/__tests__/empty_repo_list-test.jsx
@@ -1,4 +1,4 @@
-import EmptyRepoList from '../empty_repo_list.jsx';
+import EmptyRepoList from '../components/ReposContainer/components/Organization/RepoList.jsx';
 
 it('renders an empty unordered list', () => {
   const wrapper = shallow(
