+++ app/javascript/__tests__/organization-test.jsx
@@ -1,4 +1,4 @@
-import Organization from '../organization.jsx';
+import Organization from '../components/ReposContainer/components/Organization.jsx';
 
 it('renders an organization with ID appropriately (no repo processing)', () => {
   const org = {
