+++ app/javascript/__tests__/repo_tools-test.jsx
@@ -1,4 +1,4 @@
-import RepoTools from '../repo_tools.jsx';
+import RepoTools from '../components/ReposContainer/components/RepoTools';
 
 it('renders appropriately without Show Private button (not syncing)', () => {
   const has_private_access = true;
